created post-body component
